pyproject.toml
[tool.poetry]
name = "gemini-structured"
version = "0.1.0"
description = ""
authors = ["tripercy <percytri37@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.7.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
